Use long for service id
git-svn-id: https://svn.apache.org/repos/asf/sling/trunk/bundles/api@1690821 13f79535-47bb-0310-9956-ffa450edef68

+++ src/main/java/org/apache/sling/api/resource/runtime/dto/ResourceProviderDTO.java
@@ -52,5 +52,5 @@ public class ResourceProviderDTO {
     /**
      * The service id from the service registry.
      */
-    public int serviceId;
+    public long serviceId;
 }
